# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Discord Music Bot API
  version: 1.0.0
  description: |
    RESTful API for the Discord Music Bot system.

    Provides endpoints for:
    - Analytics and metrics
    - Guild management
    - Music queue operations
    - Premium subscriptions
    - Webhook integrations (Stripe, Discord)

    ## Authentication
    All endpoints require API key authentication via the `X-API-Key` header,
    except for webhook endpoints which use signature verification.

    ## Rate Limiting
    - General endpoints: 100 requests/minute
    - Search endpoints: 10 requests/minute
    - Webhook endpoints: 50 requests/minute

  contact:
    name: Discord Bot Team
    url: https://github.com/your-org/discord-bot
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.discordmusicbot.com/api/v1
    description: Production server (global)
  - url: https://staging-api.discordmusicbot.com/api/v1
    description: Pre-production / staging server
  - url: https://sandbox-api.discordmusicbot.com/api/v1
    description: Internal development sandbox (VPN only)
...
security:
  - ApiKeyAuth: []
paths:
...
  /player/{guildId}/now-playing:
    get:
      tags:
        - Music
      summary: Get live playback state
      description: Returns the current playback state reported by the audio service cache
      operationId: getNowPlayingState
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Playback state retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NowPlaying'
        '404':
          description: Guild not found or no playback data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /player/{guildId}/events:
    get:
      tags:
        - Music
      summary: Subscribe to live playback events
      description: |-
        Server-Sent Events (SSE) stream that emits `snapshot` and `update` payloads when the audio service publishes
        `discord-bot:ui:now`. Clients should include the API key via query string (`?apiKey=`) and keep the connection alive.
      operationId: subscribePlayerEvents
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Event stream opened
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /player/{guildId}/stream-token:
    post:
      tags:
        - Music
      summary: Generate browser playback token
      description: Issues a short-lived token that authorizes `/player/{guildId}/stream`.
      operationId: createPlayerStreamToken
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      responses:
        '200':
          description: Token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamToken'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /player/{guildId}/stream:
    get:
      tags:
        - Music
      summary: Proxy audio stream for panel playback
      description: |-
        Streams the current track (YouTube sources are transcoded via ytdl-core; other URIs are proxied as-is). Requires a
        valid `token` query parameter obtained from `/player/{guildId}/stream-token`.
      operationId: streamPlayerAudio
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audio stream
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /player/{guildId}/controls:
    post:
      tags:
        - Music
      summary: Send remote playback action
      description: Issues a playback action (toggle, skip, stop, shuffle, mute, volume) against the active guild player
      operationId: sendPlayerControl
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerControlRequest'
      responses:
        '200':
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerControlResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /player/{guildId}/summon:
    post:
      tags:
        - Music
      summary: Premium bot summon from panel
      description: Invokes the Discord bot into a selected voice channel (premium guilds only).
      operationId: summonBotFromPanel
      parameters:
        - $ref: '#/components/parameters/GuildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummonRequest'
      responses:
        '200':
          description: Summon accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerControlResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    GuildIdParam:
      name: guildId
      in: path
      description: Discord guild ID
      required: true
      schema:
        type: string
        pattern: '^\d{17,19}$'
...
  responses:
    BadRequestError:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
...
    SummonRequest:
      type: object
      required:
        - voiceChannelId
        - textChannelId
      properties:
        voiceChannelId:
          type: string
        textChannelId:
          type: string
...
    PlayerControlRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [toggle, skip, stop, shuffle, clear, previous, mute, volume]
        value:
          type: integer
          minimum: 0
          maximum: 200
          description: Required when action is volume
    PlayerControlResponse:
      type: object
      properties:
        accepted:
          type: boolean
    StreamToken:
      type: object
      properties:
        token:
          type: string
        expires:
          type: integer
          format: int64
    NowPlaying:
      type: object
      properties:
        guildId:
          type: string
        title:
          type: string
        author:
          type: string
          nullable: true
        uri:
          type: string
          nullable: true
        artworkUrl:
          type: string
          nullable: true
        durationMs:
          type: integer
        positionMs:
          type: integer
        isStream:
          type: boolean
        paused:
          type: boolean
        repeatMode:
          type: string
          enum: [off, track, queue]
        queueLen:
          type: integer
        hasTrack:
          type: boolean
        canSeek:
          type: boolean
        volume:
          type: integer
          minimum: 0
          maximum: 200
        autoplay:
          type: boolean
        autoplayMode:
          type: string
          enum: [off, similar, artist, genre, mixed]
        filter:
          type: object
          properties:
            id:
              type: string
            label:
              type: string
            description:
              type: string
          nullable: true
        updatedAt:
          type: integer
          format: int64
          description: Epoch milliseconds when the cache entry was produced
...